It's been, like, 20 years. Rename files.
Since Pawn was "SmallC" it's used a numbering scheme for its source
files, roughly numbered after stages of compilation:

 1 - Parser entry point, statement parsing.
 2 - Lexing.
 3 - Expression parsing.
 4 - Code generation.
 5 - Error reporting.
 6 - Assembly.
 7 - Peephole optimization.

While Pawn "veterans" know this numbering by now, it has absolutely no
meaning to anyone else. So let's call these files what they are.
