+++ compiler/AMBuilder
@@ -56,13 +56,13 @@ binary.sources += [
   'lstring.cpp',
   'memfile.cpp',
   'pawncc.cpp',
-  'sc1.cpp',
-  'sc2.cpp',
-  'sc3.cpp',
-  'sc4.cpp',
-  'sc5.cpp',
-  'sc6.cpp',
-  'sc7.cpp',
+  'parser.cpp',
+  'lexer.cpp',
+  'expressions.cpp',
+  'codegen.cpp',
+  'errors.cpp',
+  'assembler.cpp',
+  'optimizer.cpp',
   'sci18n.cpp',
   'sclist.cpp',
   'scmemfil.cpp',
@@ -70,7 +70,7 @@ binary.sources += [
   'scvars.cpp',
   'sp_symhash.cpp',
   'types.cpp',
-  'expression-parsing.cpp',
+  'parse-expr.cpp',
 ]
 
 if builder.target.platform == 'linux' and not compiler.like('emscripten'):
