+++ compiler/assembler.cpp
