+++ compiler/parse-expr.cpp
@@ -21,7 +21,7 @@
  *
  *  Version: $Id$
  */
-#include "expression-parsing.h"
+#include "parse-expr.h"
 #include "sc.h"
 #include "types.h"
 #include <assert.h>
