+++ compiler/patterns.h
